move check to alpha module
